pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo-inspector"
version = "0.1.0"
description = "Understand any codebase in minutes"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "repo-inspector contributors" }
]
keywords = ["cli", "codebase", "analysis", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "google-genai>=1.0.0",
]

[project.scripts]
repo-inspector = "repo_inspector.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["repo_inspector"]
